fixed intended bug
